# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for live stream configuration, ingest/publish lifecycle, DVR recordings, and event hooks.",
        "title": "Open Streamer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/config": {
            "get": {
                "description": "Returns available hardware accelerators (OS-detected), static enum lists, publisher listener ports, and the current runtime GlobalConfig.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get server configuration.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ConfigData"
                        }
                    }
                }
            },
            "post": {
                "description": "Merges the request body into the current GlobalConfig and applies it — services are started or stopped to match.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Partially update server configuration.",
                "parameters": [
                    {
                        "description": "Partial configuration to merge",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.GlobalConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ConfigUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/config/transcoder/probe": {
            "post": {
                "description": "Inspects the binary at ffmpeg_path (empty = $PATH) and reports which required / optional encoders + muxers are available. Pure check — does not modify config.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Probe an FFmpeg binary for app compatibility.",
                "parameters": [
                    {
                        "description": "Path to probe (empty = use $PATH)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.probeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transcoder.ProbeResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/config/yaml": {
            "get": {
                "description": "Returns global_config, streams, and hooks bundled in a single YAML document. Pair with PUT /config/yaml to round-trip an editor.",
                "produces": [
                    "application/yaml"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get full system configuration as YAML",
                "responses": {
                    "200": {
                        "description": "YAML document",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Replaces global_config, streams, and hooks with the request body. Validation errors are returned as a list. On success, the runtime manager + coordinator diff against the previous state and start/stop/restart services accordingly.",
                "consumes": [
                    "application/yaml"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Replace full system configuration from YAML",
                "parameters": [
                    {
                        "description": "Full system configuration YAML document",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/hooks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "List hooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Create hook",
                "parameters": [
                    {
                        "description": "Hook configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Hook"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/hooks/{hid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Get hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Update hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hook configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Hook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "hooks"
                ],
                "summary": "Delete hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/recordings/{rid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "Get recording",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recording ID (= stream code)",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.RecordingData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/recordings/{rid}/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "DVR recording info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recording ID (= stream code)",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/sessions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "List play sessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by protocol (hls|dash|rtmp|srt|rtsp)",
                        "name": "proto",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active|closed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max sessions to return (default 0 = no cap)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.SessionList"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apidocs.ConfigData": {
            "type": "object",
            "properties": {
                "audio_codecs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AudioCodec"
                    }
                },
                "global_config": {
                    "$ref": "#/definitions/domain.GlobalConfig"
                },
                "hw_accels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.HWAccel"
                    }
                },
                "output_protocols": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ports": {
                    "$ref": "#/definitions/apidocs.ConfigPorts"
                },
                "stream_statuses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StreamStatus"
                    }
                },
                "version": {
                    "$ref": "#/definitions/version.Info"
                },
                "video_codecs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.VideoCodec"
                    }
                },
                "watermark_positions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WatermarkPosition"
                    }
                },
                "watermark_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WatermarkType"
                    }
                }
            }
        },
...
        "handler.probeRequest": {
            "type": "object",
            "properties": {
                "ffmpeg_path": {
                    "type": "string"
                }
            }
        },
...
    }
}